Move Zernike tests to test_zernike.py
